Move is_input_and_output_consistent to Transaction
