+++ types/src/transaction/error.rs
@@ -46,6 +46,8 @@ pub enum Error {
     /// Script execution result is `Fail`
     FailedToUnlock(H256),
     InvalidNetworkId(Mismatch<u64>),
+    /// Returned when the sum of the transaction's inputs is different from the sum of outputs.
+    InconsistentTransactionInOut,
 }
 
 impl Display for Error {
@@ -76,6 +78,9 @@ impl Display for Error {
             Error::InvalidScript => write!(f, "Failed to decode script"),
             Error::FailedToUnlock(hash) => write!(f, "Failed to unlock asset {}", hash),
             Error::InvalidNetworkId(mismatch) => write!(f, "Invalid network id. {}", mismatch),
+            Error::InconsistentTransactionInOut => {
+                write!(f, "The sum of the transaction's inputs is different from the sum of the transaction's outputs")
+            }
         }
     }
 }
